test: deduplicate integration tests
